Handle pivot clauses in relationship joiner
