+++ packages/tables/src/Filters/Concerns/HasRelationship.php
@@ -3,10 +3,9 @@
 namespace Filament\Tables\Filters\Concerns;
 
 use Closure;
+use Filament\Support\Services\RelationshipJoiner;
 use Illuminate\Database\Eloquent\Builder;
-use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 use Illuminate\Database\Eloquent\Relations\Relation;
-use Illuminate\Database\Query\JoinClause;
 
 trait HasRelationship
 {
@@ -79,23 +78,7 @@ trait HasRelationship
     {
         $relationship = Relation::noConstraints(fn () => $this->getRelationship());
 
-        $relationshipQuery = $relationship->getQuery();
-
-        // By default, `BelongsToMany` relationships use an inner join to scope the results to only
-        // those that are attached in the pivot table. We need to change this to a left join so
-        // that we can still get results when the relationship is not attached to the record.
-        if ($relationship instanceof BelongsToMany) {
-            /** @var ?JoinClause $firstRelationshipJoinClause */
-            $firstRelationshipJoinClause = $relationshipQuery->getQuery()->joins[0] ?? null;
-
-            if ($firstRelationshipJoinClause) {
-                $firstRelationshipJoinClause->type = 'left';
-            }
-
-            $relationshipQuery
-                ->distinct() // Ensure that results are unique when fetching options and indicating.
-                ->select($relationshipQuery->getModel()->getTable() . '.*');
-        }
+        $relationshipQuery = (new RelationshipJoiner())->prepareQueryForNoConstraints($relationship);
 
         if ($this->getModifyRelationshipQueryUsing()) {
             $relationshipQuery = $this->evaluate($this->modifyRelationshipQueryUsing, [
